added boot strap.
this is my current code for bootstrap add ons. i got all functions of my bootsrtrap to function perfectly eccept for the bootstrap modal. all info is loding into the html. but i cant seem to get it to actually display. i commented out my old modal to keep it there as referance to what was working before.